Add package for Tessera integration
This change adds the workflow to instantiate the tessera storage driver
add an entry to the log. It copies most of the sigstore/rekor/pkg/signer
package to enable signing checkpoints with an on-disk key file or a KMS
provider.

The TransparencyLogEntry structure uses signed integers for the LogIndex
and TreeSize values, while tessera mostly returns unsigned integers for
these values. This change incorporates a conversion mechanism. It would
probably be better to update the protobuf-specs definition to use
unsigned integers.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>