Add rekor and config steps to docker compose
Update the rekor service in compose.yml to reflect the new configuration
options for the rekor service. Add a service to run some gcloud steps to
set up the Spanner emulator database instance.

This also changes the GCS emulator to run on another port to avoid
conflicting with the Rekor HTTP unit tests which start a server on the
same port.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>