Codecov tweaks (#107)
* workflows: Fail coverage upload if action fails

I have no idea why this would default to false but it seems like it
should not.

Also allow workflow dispatch for test.yml: this is useful when debugging
issues.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* workflows: Enable oidc auth for codecov

This should just work with "use_oidc: true" but does not.
Add a workaround that does the authentication for codecov action

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* workflows: Add url to upstream issue for workaround

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>