Expose Tessera append lifecycle options
Expose Tessera's batch size, batch age, checkpoint interval, and
pushback options through Rekor's configuration. This will enable us to
tune the server to balance the cost and latency of sequencing entries,
and in particular it will allow us to reduce the time to publish a new
checkpoint so that entries added sequentially in tests no longer need to
wait up to 10 seconds for a response.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>