Add entry hash algorithm config and validation
Port over the configurable hash algorithms and key validation from
sigstore/rekor PR 1974. This both introduces a flag to restrict the
allowed signing algorithms for entries being submitted, as well as
validation to ensure the key used is within established constrictions
for the specified algorithm.

Much of this was borrowed from Rekor v1 with some modifications to
account for the different structure in rekor-tiles. Testing was done
here as unit tests rather than porting the script for the integration
tests from that PR.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>