Add support for deduplication
Add options to configure the WithAntispam Tessera append options[1].

The "persistent" antispam layer does not work with the GCP Spanner
emulator because it relies on the currently unsupported "batch write"
operation. So this option is disabled by default and not tested in the
e2e tests, but should be enabled when deployed to real infrastructure.

[1] https://github.com/transparency-dev/trillian-tessera?tab=readme-ov-file#antispam

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>