Fix metrics test failure when run by itself
Fixes #162

This failure occurs because the HTTP latency and request metrics seem to
be created only when a request has occurred.

When running all tests with go test ./..., no failure occurs, which is
why CI didn't catch this. I believe this succeeded because another test
triggered an HTTP request and however Prometheus is scraping metrics
created shared state between tests.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>