Add support for freezing the log
Add a flag that, when Rekor is restarted, prevents the service from
accepting new entries or publishing new checkpoints. Add a command to
publish a final checkpoint that contains an extension indicating the log
is frozen.

Technically Rekor could be fully shut down and it would accomplish the
same purpose of not accepting new entries or updating new checkpoints.
The server change is mostly useful in the potential future when Rekor
needs to be responsible for serving checkpoints stored in MySQL. The
freeze command is necessary no matter the backend.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>