Shutdown Tessera gracefully on SIGTERM/SIGINT (#161)
Fixes #138

When SIGTERM/SIGINT is sent to the process, which occurs when Kubernetes
is shutting down pods, the server gracefully shuts down the gRPC and
HTTP servers. Tessera also provides a shutdown function which will wait
until a checkpoint has been published that covers all sequenced entries.

This change plumbs this shutdown function through to call it when the
server is shutting down.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>