Refactor pkg/pki into verifier library
I've redone the pkg/pki library to greatly simplify it, beacuse at this
point, we have no path forward for supporting any other verifiers
besides certificates and public keys due to our bundle format. The
changes are roughly as follows:

* Renamed PKI interface to Verifier, and changed method names
* Split x509 verifier into two, for public keys and for certificates
* Removed the Signature interface. DSSE doesn't use it because it relies
  on a sigstore/sigstore DSSE verification library, so it was simpler to
  just drop it and use the sig verification methods directly in
  hashedrekord
* Renamed existing verifier package to validator to clarify difference

I also fixed a bug that assumed the verifier input was PEM-encoded. The
protobuf spec states that the certificate and public key RawBytes are
DER-encoded.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>