Specify correct hash algorithm in response
This fixes a bug for hashedrekord where the response hash algorithm was
always hardcoded to SHA-256. For DSSE, the fix was a bit more tricky.
Because we may have multiple PublicKeyDetails with different hash
algorithms, because a DSSE can have multiple signatures, we need to
select one of the algorithms to use to hash the payload.

We could have hardcoded SHA-256, but this is a step backwards for crypto
agility. We could reject requests with different hash algorithms, but
this would prevent artifacts signed with classical and PQ crypto
algorithms. We could outright reject DSSEs with multiple sigs, but this
is likely to have a larger impact. Therefore, we implemented a simple
selection algorithm to pick the strongest and newest digest alg.

Also fixes a bug where the order of the signatures was not consistent
when multiple signatures are returned in the DSSE response.

Fixes #195

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>