Refactor ToLogEntry, extract validators and verifiers
Cleans up ToLogEntry for hashedrekord and dsse to try to make it more
explicit that the function is doing both request validation and
verificaction of entry signatures. Hashedrekord refactored very easily,
but DSSE is a bit more tricky given the repeated verifiers. May come
back to this later.

For now, refactored extracted functions remain unexported as an external
use case is not yet known.

Fixes #178

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>