Change condition for building tagged container
The `event_name` is `push`, not `tag`. Since the only event to trigger this GHA is `push`, we can omit that. To differentiate between branch pushes and tags, we can use `github.ref`, which will either be `refs/heads/main` when pushed to main or `refs/tags/v1.2.3` for tags.

Signed-off-by: Hayden B <haydentherapper@users.noreply.github.com>