Include KindVersion in TLE response
This adds a missing required field for the kind and version of the
entry, so that clients will know how to parse the canonicalized body.

This is a much simpler approach to the type factory in Rekor. At this
point, given there are only two types, I don't think we should
reimplement the factory and simply use case statements.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>