Include type and version of entry for client verification
We have decided to mirror the Rekor v1 response to include API version,
kind, and spec, and clients should use the former two to parse the spec.
Clients should be able to handle a kind and version that they don't
support and gracefully fail.

Note that kind and version is required in the entry in addition to the
bundle. The type information is used by monitors parsing entries from
the log, and within the bundle is used by clients that do not have the
canonicalized body.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>