Use Go cross-compiler rather than emulated build platform
Fixes #229

Following this example:
https://docs.docker.com/build/building/multi-platform/#cross-compiling-a-go-application

Rather than use QEMU to launch an ARM build environment, which is
incredibly slow on GitHub Actions, this uses Go's native cross
compilation to build for multiple platforms. The builder is pinned to
the build platform, Linux, while the target platform is passed to the Go
builder as an OS and architecture.

Ko wasn't used because it doesn't build a multi-platform image, it
builds an image per platform.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>