Build dlv with cross-compilation
The build time is now down from 34 min to 10 min, but we were still
installing dlv on a build platform that didn't match the runner
platform. This change reuses the builder image, whose platform matches
the runner, to cross-compile dlv, and then copy the binary over in a
later stage.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>