Fix cross-compilation for dlv binary for Docker
When the build and target platform are the same, go install installs the
binary under GOPATH/bin. If they differ, then it is installed under
GOPATH/bin/TARGETOS_TARGETARCH. I didn't notice this in my local
testing, since the build and target platforms of course matched. In CI,
this caused an issue when the build platform (linux_amd64) did not match
the target platform (linux_arm64), and the binary was installed in an
unexpected directory.

Note that go install doesn't let you select output location, and go
build doesn't let you target specific versions. There are feature
requests for both of these.

Also copied in the ENV variables into the dlvbuilder stage, since ARG
and ENV steps are not copied into later stages.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>