Fix root hash as bytes rather than hex-encoded
To produce strings in Rekor v1, all hashes were hex-encoded. When the
protobuf specs were created, rather than specify strings for hashes, we
specified bytes. This fixes the root hash of the inclusion proof to be a
byte array (and when marshalled to JSON, it will be base64-encoded).

This is yet another example of what should be updated with a TLE v2,
dropping log index (specified in the bundle), root hash and tree size
(both specified in the checkpoint). I'll update the client spec to
recommend ignoring these fields.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>