Remove verification from upload method
Without this change, the client support package requires a public key
verifier and an origin string to verify entries from the transparency
log immediately after they are uploaded. This requires clients to have a
trusted root for signing, and requires them to deduce from the signing
config and the trusted root which key in the key material is correct for
that server. This is unnecessarily cumbersome, when a client could
easily verify the log entry separately using its trusted root and the
log ID returned in the server response. This change removes the
requirement for a verifier for uploading an entry to avoid this
unnecessary complication.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>