Change default origin for test setup
Clients need to deduce the checkpoint signer's origin in order to verify
entries, and have no good way to accept a hint if the origin string does
not match the server hostname the client is connecting to. This change
updates the default hostname in the docker compose setup to match what
most tests or local deployments will usually use to connect to Rekor.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>