Create initial sharding playbook (#254)
This includes the sharding playbook to be used by oncallers to spin up a
new shard.

There's a TODO to add a dashboard link once we have monitoring, and WIPs
as we are actively developing the prober and figuring out shard turn
down.

Towards https://github.com/sigstore/rekor-tiles/issues/40

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>