Add better shutdown handling for tessera (#432)
Before this change. it was possible for Tessera to get in the way of the
process shutting down in two ways:

1) The context in the Appender did not respond to signals, so if it ran
   into an irrecoverable error in `integrateEntriesJob`, it would retry
   indefinitely and the process would not end shutdown.
2) The appender Shutdown function also runs in an indefinite loop, and
   if Tessera got into a corrupt state where it could never finish
   appending the sequenced entries, the shutdown loop would never
   terminate.

This change adds a new timeout parameter specifically for Tessera to
give it a maximum time to finish processing entries. In the event it is
stuck looping trying to process an unprocessable entry, it will now
terminate in response to SIGINT/SIGTERM.

The default timeout for the server timeout, which is shared by the HTTP
and gRPC servers, is reduced to 20s, which in a correctly functioning
system with a checkpoint period of 10s should be long enough to finish
processing all in-flight requests. The new tessera timeout is 30s. This
means that in a badly behaving system where tessera did not integrate
entries in time to send the response to clients, it still has 30s to try
to finish committing entries to the log and empty the Spanner queue.

In the event that tessera encounters an irrecoverable error and cannot
finish processing entries, the maximum time to shut down is 50s by
default. This is reduced from 60s (server idle timeout) + indefinite
(tessera loop). The Pod termination grace period should be extended to
50-60s accordingly.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>