Include log index as response header when duplicate is detected
If a duplicate entry is detected during a request, the server returns a
409 with an error message that contains the log index. To avoid clients
needing to know the structure of this error message, the server will now
return a custom header "x-log-index" with the log index value. A client
can use this log index to retrieve the entry from the tiled entry bundle
and compute an inclusion proof from the tiles.

An alternative approach to solving this would be to just not return 409
errors and instead return the duplicate content. This would mask whether
or not an entry is new content. This might be acceptable to clients, but
to give them choice on how to respond to duplicates, we'll continue to
serve 409 errors.

Fixes https://github.com/sigstore/rekor-tiles/issues/620

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>